+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * **New** - Elementor Dynamic Tag: Criteo Product ID
+* **New** - ACF for Discovery: new field video
 
 ### 1.1.7: September 20th, 2018
 * **New** - Elementor Dynamic Tag: Resaweb URL No Accommodation